requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bytesbufio==1.0.3
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
colorama==0.4.5
contourpy==1.1.0
cryptography==38.0.1
cycler==0.10.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.0.6
django-allauth==0.51.0
django-bootstrap4==22.1
django-environ==0.10.0
django-extensions==3.2.1
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
executing==1.1.0
filelock==3.8.0
fonttools==4.41.0
gunicorn==21.2.0
h11==0.14.0
idna==2.10
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
linecache2==1.0.0
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.11.5
platformdirs==2.5.3
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
pyOpenSSL==22.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0.1
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.1
selenium==3.14.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.1
supervision==0.11.1
tornado==6.2
tqdm==4.65.0
traceback2==1.4.0
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
unittest2==1.1.0
urllib3==1.26.10
virtualenv==20.16.6
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.2.2
wget==3.2
whitenoise==6.2.0
wsproto==1.2.0
